requirements.txt
PySide6>=6.5.0
PyPDF2>=3.0.0
pypdf>=3.15.0
pdfkit>=1.0.0
markdown>=3.4.0
pypandoc>=1.10.0
Pillow>=9.0.0
requests>=2.28.0
reportlab>=3.6.0